refactor(handler): extract shared session cookie helper
